9. Редактирование и удаление только своих заметок

+++ notes.controller.js
@@ -13,13 +13,25 @@ async function getNotes() {
   return notes;
 }
 
-async function removeNote(id) {
-  await Note.deleteOne({ _id: id });
+async function removeNote(id, owner) {
+  const result = await Note.deleteOne({ _id: id, owner });
+
+  if (result.matchedCount === 0) {
+    throw new Error('No note to delete');
+  }
+
   console.log(chalk.red(`Note with id="${id}" has been removed.`));
 }
 
-async function updateNote(noteData) {
-  await Note.updateOne({ _id: noteData.id }, { title: noteData.title });
+async function updateNote(noteData, owner) {
+  const result = await Note.updateOne(
+    { _id: noteData.id, owner },
+    { title: noteData.title }
+  );
+  console.log('result', result);
+  if (result.matchedCount === 0) {
+    throw new Error('No note to edit');
+  }
   console.log(chalk.bgGreen(`Note with id="${noteData.id}" has been updated!`));
 }
